+++ tests/testthat/test-compat.R
@@ -6,7 +6,7 @@ test_that("names() dispatches on environment", {
 })
 
 test_that("lazy objects are converted to tidy quotes", {
-  env <- child_env(get_env())
+  env <- child_env(current_env())
 
   lazy <- structure(list(expr = quote(foo(bar)), env = env), class = "lazy")
   expect_identical(compat_lazy(lazy), new_quosure(quote(foo(bar)), env))
@@ -22,7 +22,7 @@ test_that("lazy objects are converted to tidy quotes", {
 })
 
 test_that("lazy_dots objects are converted to tidy quotes", {
-  env <- child_env(get_env())
+  env <- child_env(current_env())
 
   lazy_dots <- structure(class = "lazy_dots", list(
     lazy = structure(list(expr = quote(foo(bar)), env = env), class = "lazy"),
@@ -35,7 +35,7 @@ test_that("lazy_dots objects are converted to tidy quotes", {
     quo(foo(bar))
   )
 
-  expect_identical(compat_lazy_dots(lazy_dots, get_env(), "foo(bar)"), expected)
+  expect_identical(compat_lazy_dots(lazy_dots, current_env(), "foo(bar)"), expected)
 })
 
 test_that("unnamed lazy_dots are given default names", {
@@ -49,7 +49,7 @@ test_that("unnamed lazy_dots are given default names", {
     `foo(bar)` = quo(foo(bar)),
     foobarbaz = quo(foo(barbaz))
   )
-  dots <- compat_lazy_dots(lazy_dots, get_env(), foobarbaz = "foo(barbaz)", .named = TRUE)
+  dots <- compat_lazy_dots(lazy_dots, current_env(), foobarbaz = "foo(barbaz)", .named = TRUE)
 
   expect_identical(dots, expected)
 })
@@ -69,12 +69,12 @@ test_that("compat_lazy_dots() takes symbolic objects", {
 })
 
 test_that("compat_lazy() demotes character vectors to strings", {
-  expect_identical(compat_lazy_dots(NULL, get_env(), c("foo", "bar")), named_list(as_quosure(~foo)))
+  expect_identical(compat_lazy_dots(NULL, current_env(), c("foo", "bar")), named_list(as_quosure(~foo)))
 })
 
 test_that("compat_lazy() handles numeric vectors", {
-  expect_identical(compat_lazy_dots(NULL, get_env(), NA_real_), named_list(set_env(quo(NA_real_))))
-  expect_warning(expect_identical(compat_lazy_dots(NULL, get_env(), 1:3), named_list(set_env(quo(1L)))), "Truncating vector")
+  expect_identical(compat_lazy_dots(NULL, current_env(), NA_real_), named_list(set_env(quo(NA_real_))))
+  expect_warning(expect_identical(compat_lazy_dots(NULL, current_env(), 1:3), named_list(set_env(quo(1L)))), "Truncating vector")
 })
 
 test_that("compat_lazy() handles bare formulas", {